restructured my-environment and common services
