+++ src/app/features/my-environment/pages/my-environment-result/my-environment-result.component.html
@@ -1,63 +1,56 @@
 
-
-  <mat-grid-list [cols]="breakpoint" rowHeight="550px">
-    <mat-grid-tile>
-      <div class="container">
-        <div class=" results-card">
-          <babcn-container containerHeight="395px" [containerWidth]='ratio.toString()' vaStyle="theme">
-          <div class="bg-light">
-    <h2 >
-      
-      {{ "common.title.results" | translate }}
-    </h2>
-    <div class="data" *ngIf="true" id="pdfData">
-      <div
-      *ngFor="let ls of businessModelsArray"
-      >
-      <h5 class="details">{{ ls.name }}</h5>
-      <mat-list class="result-container list-unstyled" *ngFor="let address of ls.addresses">
-        <!--*ngFor must be added!-->
-        <mat-list-item *ngIf="address.street_name">
-          {{ address.street_name }}, {{ address.street_number }}
-        </mat-list-item>
-        <mat-list-item *ngIf="address.town">
-          {{ address.zip_code }} {{ address.town }}
-        </mat-list-item>
-        <mat-list-item *ngIf="ls.email">Email: {{ ls.email }}</mat-list-item>
-        <mat-list-item *ngIf="ls.web">Web: {{ ls.web }}</mat-list-item>
-        <mat-list-item *ngIf="ls.phone">Tel.:{{ ls.phone }}</mat-list-item>
-        <mat-list-item></mat-list-item>
-      </mat-list>
-      </div>
-    </div>
-    <button mat-raised-button color="accent" style="background: #2c5e8f; color: white">
-    {{ "common.button.save-results" | translate }}
-    </button>
-    <button mat-raised-button color="accent" style="background: #5cb85c; color: white"
-     id="dropdownMenuButton1" data-bs-toggle="dropdown" aria-expanded="false">
-    {{ "common.button.download-results" | translate }}
-    </button>
-
-
-        </div>
-      </babcn-container>
-      </div>
-    </div>
-    </mat-grid-tile>
+<mat-grid-list [cols]="breakpoint" rowHeight="550px">
 
     <mat-grid-tile>
       <div class="container">
-        
-        <div class=" results-card">
           <babcn-container containerHeight="395px" [containerWidth]='ratio.toString()' vaStyle="theme">
-          <div class="bg-light">
-            <app-mapbox [filteredResultsToPrintOnMap]="results"></app-mapbox>
-        </div>
-        </babcn-container>
+            <h2 >
+              {{ "common.title.results" | translate }}
+            </h2>
+            <div class="data" *ngIf="true" id="pdfData">
+              <div
+              *ngFor="let ls of businessModelsArray"
+              >
+              <h5 class="details">{{ ls.name }}</h5>
+              <mat-list class="result-container list-unstyled" *ngFor="let address of ls.addresses">
+                <mat-list-item *ngIf="address.street_name">
+                  {{ address.street_name }}, {{ address.street_number }}
+                </mat-list-item>
+                <mat-list-item *ngIf="address.town">
+                  {{ address.zip_code }} {{ address.town }}
+                </mat-list-item>
+                <mat-list-item *ngIf="ls.email">
+                  Email: {{ ls.email }}
+                </mat-list-item>
+                <mat-list-item *ngIf="ls.web">
+                  Web: {{ ls.web }}
+                </mat-list-item>
+                <mat-list-item *ngIf="ls.phone">
+                  Tel.:{{ ls.phone }}
+                </mat-list-item>
+              </mat-list>
+              </div>
+            </div>
+            <button mat-raised-button color="accent" style="background: #2c5e8f; color: white">
+            {{ "common.button.save-results" | translate }}
+            </button>
+            <button mat-raised-button color="accent" style="background: #5cb85c; color: white"
+            id="dropdownMenuButton1" data-bs-toggle="dropdown" aria-expanded="false">
+            {{ "common.button.download-results" | translate }}
+            </button>
+          </babcn-container>
       </div>
-    </div>
     </mat-grid-tile>
-  </mat-grid-list>
+
+    <mat-grid-tile>
+        <div class="container"> 
+            <babcn-container containerHeight="395px" [containerWidth]='ratio.toString()' vaStyle="theme">
+              <app-mapbox [filteredResultsToPrintOnMap]="results"></app-mapbox>
+            </babcn-container>
+        </div>
+    </mat-grid-tile>
+    
+</mat-grid-list>
