+++ src/app/features/my-environment/pages/my-environment-result/my-environment-result.component.scss
@@ -9,10 +9,6 @@
   .details{
     padding-left: 12%;
   }
-
-  .bg-light {
-    background-color: white;
-  }
   
   .data{
     overflow-y: scroll;
